raylib: refactor NetworkManager constants (#36056)
* new file

* import

* and this
