+++ system/ui/lib/wifi_manager.py
@@ -14,6 +14,11 @@ from dbus_next import BusType, Variant, Message
 from dbus_next.errors import DBusError
 from dbus_next.constants import MessageType
 
+from openpilot.system.ui.lib.networkmanager import (NM, NM_PATH, NM_IFACE, NM_SETTINGS_PATH, NM_SETTINGS_IFACE,
+                                                    NM_CONNECTION_IFACE, NM_WIRELESS_IFACE, NM_PROPERTIES_IFACE,
+                                                    NM_DEVICE_IFACE, NM_DEVICE_STATE_REASON_SUPPLICANT_DISCONNECT,
+                                                    NMDeviceState)
+
 try:
   from openpilot.common.params import Params
 except ImportError:
@@ -23,30 +28,8 @@ from openpilot.common.swaglog import cloudlog
 
 T = TypeVar("T")
 
-# NetworkManager constants
-NM = "org.freedesktop.NetworkManager"
-NM_PATH = '/org/freedesktop/NetworkManager'
-NM_IFACE = 'org.freedesktop.NetworkManager'
-NM_SETTINGS_PATH = '/org/freedesktop/NetworkManager/Settings'
-NM_SETTINGS_IFACE = 'org.freedesktop.NetworkManager.Settings'
-NM_CONNECTION_IFACE = 'org.freedesktop.NetworkManager.Settings.Connection'
-NM_WIRELESS_IFACE = 'org.freedesktop.NetworkManager.Device.Wireless'
-NM_PROPERTIES_IFACE = 'org.freedesktop.DBus.Properties'
-NM_DEVICE_IFACE = "org.freedesktop.NetworkManager.Device"
-
-NM_DEVICE_STATE_REASON_SUPPLICANT_DISCONNECT = 8
-
 TETHERING_IP_ADDRESS = "192.168.43.1"
 DEFAULT_TETHERING_PASSWORD = "swagswagcomma"
-
-
-# NetworkManager device states
-class NMDeviceState(IntEnum):
-  DISCONNECTED = 30
-  PREPARE = 40
-  NEED_AUTH = 60
-  IP_CONFIG = 70
-  ACTIVATED = 100
 
 
 class SecurityType(IntEnum):
